requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.1.0
pydantic==2.5.2
openai==1.86.0
python-dotenv==1.0.0 